+++ ochre/Dwelling.py
@@ -11,7 +11,7 @@ import pandas as pd
 import numpy as np
 
 from ochre import Simulator, Analysis
-from ochre.utils import load_hpxml, load_schedule, nested_update, update_equipment_properties, save_json
+from ochre.utils import OCHREException, load_hpxml, load_schedule, nested_update, update_equipment_properties, save_json
 from ochre.Models import Envelope
 from ochre.Equipment import *
 
@@ -27,6 +27,7 @@ class Dwelling(Simulator):
                  **house_args):
         super().__init__(**house_args)
         house_args.pop('name', None)  # remove name from kwargs
+        house_args['main_sim_name'] = self.name
 
         # Time parameters
         if self.initialization_time is not None:
@@ -45,21 +46,23 @@ class Dwelling(Simulator):
         self.metrics_verbosity = metrics_verbosity
         _ = house_args.pop('save_results', None)  # remove save_results from args to prevent saving all Equipment files
         if self.output_path is not None:
-            # remove existing output files, and save file locations
+            # remove existing output files
+            for file_type in ['metrics', 'hourly', 'schedule']:
+                for extn in ['parquet', 'csv']:
+                    f = os.path.join(self.output_path, f'{self.name}_{file_type}.{extn}')
+                    if os.path.exists(f):
+                        self.print('Removing previous results file:', f)
+                        os.remove(f)
+
+            # save file locations
             extn = '.parquet' if self.output_to_parquet else '.csv'
             self.metrics_file = os.path.join(self.output_path, self.name + '_metrics.csv')
-            if os.path.exists(self.metrics_file):
-                os.remove(self.metrics_file)
             if self.verbosity >= 3:
                 self.hourly_output_file = os.path.join(self.output_path, self.name + '_hourly' + extn)
-                if os.path.exists(self.hourly_output_file):
-                    os.remove(self.hourly_output_file)
             else:
                 self.hourly_output_file = None
             if self.verbosity >= 7 or save_schedule_columns:
                 schedule_output_file = os.path.join(self.output_path, self.name + '_schedule' + extn)
-                if os.path.exists(schedule_output_file):
-                    os.remove(schedule_output_file)
             else:
                 schedule_output_file = None
         else:
@@ -102,7 +105,6 @@ class Dwelling(Simulator):
             'start_time': self.start_time,  # updates time zone if necessary
             'schedule': schedule,
             'initial_schedule': schedule.loc[self.start_time].to_dict(),
-            'main_sim_name': self.name,
             'output_path': self.output_path,
         }
 
@@ -134,7 +136,7 @@ class Dwelling(Simulator):
             # check if there is more than 1 equipment per end use. Raise error for HVAC/WH, else print a warning
             if len(eq) > 1:
                 if end_use in ['HVAC Heating', 'HVAC Cooling', 'Water Heating']:
-                    raise Exception(f'More than 1 equipment defined for {end_use}: {eq}')
+                    raise OCHREException(f'More than 1 equipment defined for {end_use}: {eq}')
                 elif end_use not in ['Lighting', 'Other']:
                     self.warn(f'More than 1 equipment defined for {end_use}: {eq}')
 
@@ -143,7 +145,7 @@ class Dwelling(Simulator):
             ideal = eq.use_ideal_capacity if isinstance(eq, (HVAC, WaterHeater)) else True
             if not ideal:
                 if self.time_res >= dt.timedelta(minutes=15):
-                    raise Exception(f'Cannot use non-ideal equipment {name} with large time step of'
+                    raise OCHREException(f'Cannot use non-ideal equipment {name} with large time step of'
                                              f' {self.time_res}')
                 if self.time_res >= dt.timedelta(minutes=5):
                     self.warn(f'Using non-ideal equipment {name} with large time step of {self.time_res}')
@@ -188,7 +190,7 @@ class Dwelling(Simulator):
         elif end_use in self.equipment:
             return self.equipment[end_use]
         else:
-            raise Exception(f'Unknown end use: {end_use}')
+            raise OCHREException(f'Unknown end use: {end_use}')
 
     def update_inputs(self, schedule_inputs=None):
         if schedule_inputs is None:
@@ -197,7 +199,7 @@ class Dwelling(Simulator):
         # check voltage from external model
         self.voltage = schedule_inputs.get('Voltage (-)', 1)
         if np.isnan(self.voltage) or self.voltage < 0:
-            raise Exception(f'Error reading voltage for house {self.name}: {self.voltage}')
+            raise OCHREException(f'Error reading voltage for house {self.name}: {self.voltage}')
         if self.voltage == 0:
             # Enter resilience mode when voltage is 0. Assumes home generator maintains voltage at 1 p.u.
             schedule_inputs['Voltage (-)'] = 1
@@ -257,16 +259,14 @@ class Dwelling(Simulator):
     def start_sub_update(self, sub, control_signal):
         sub_control_signal = super().start_sub_update(sub, control_signal)
         
-        if isinstance(sub, Generator):
-            # Add house net_power and pv_power to control signal for equipment that use it (Generator and Battery)
-            if sub_control_signal is None:
-                sub_control_signal = {}
+        # Add house net_power to schedule for Generator
+        if isinstance(sub, Generator) and 'net_power' not in sub.current_schedule:
+            sub.current_schedule["net_power"] = self.total_p_kw
 
-            if 'net_power' not in sub_control_signal:
-                sub_control_signal['net_power'] = self.total_p_kw
-            if isinstance(sub, Battery) and 'pv_power' not in sub_control_signal:
-                pv_power = sum([e.electric_kw for e in self.equipment_by_end_use['PV']])
-                sub_control_signal['pv_power'] = pv_power
+        # Add pv_power to schedule for Battery
+        if isinstance(sub, Battery) and "pv_power" not in sub.current_schedule:
+            pv_power = sum([e.electric_kw for e in self.equipment_by_end_use['PV']])
+            sub.current_schedule["pv_power"] = pv_power
 
         return sub_control_signal