+++ ochre/utils/base.py
@@ -7,11 +7,14 @@ import collections
 import xmltodict
 
 this_path = os.path.dirname(__file__)
-main_path = os.path.abspath(os.path.join(this_path, os.pardir, os.pardir))
+main_path = os.path.abspath(os.path.join(this_path, os.pardir))
 default_input_path = os.path.join(main_path, 'defaults')
 
 
 # TODO: add exception classes, print functions here
+
+class OCHREException(Exception):
+    pass
 
 
 def nested_update(d_old, d_new):
@@ -87,7 +90,7 @@ def convert_hpxml_element(obj, use_sys_id):
             return obj
 
     else:
-        raise Exception(f'Unknown HPXML object type ({type(obj)}: {obj}')
+        raise OCHREException(f'Unknown HPXML object type ({type(obj)}: {obj}')
 
 
 def import_hpxml(hpxml_file, use_sys_id=False, **house_args):