+++ test/__init__.py
@@ -2,6 +2,6 @@ import os
 
 from ochre.utils import main_path
 
-test_path = os.path.join(main_path, 'test')
+test_path = os.path.join(main_path, os.pardir, 'test')
 test_output_path = os.path.join(test_path, 'outputs')